move used interp files to kernel
