+++ src/main/scala/ammonite/kernel/CompilerCompatibility.scala
@@ -1,4 +1,4 @@
-package ammonite.runtime
+package ammonite.kernel
 
 import scala.tools.nsc.Global
 import scala.tools.nsc.interactive.{Global => InteractiveGlobal}
